reorganize class
