+++ src/index.ts
@@ -1,3 +1,7 @@
+import * as Locks from "./locks/lock-mutex";
+
 export function greet(name: string): string {
   return `Hello, ${name}!`;
 }
+
+export { Locks };
