--- src/interfaces/lock-entry.ts
@@ -1,10 +1,0 @@
-import { ReleaseFunction } from "./release-function";
-
-export interface WaitingLock {
-  resolve(result: ReleaseFunction): void;
-  reject(reason?: any): void;
-}
-
-export interface AcquiredLock {
-  releaseTimeoutId?: NodeJS.Timeout;
-}
