+++ src/interfaces/lock.ts
@@ -1,5 +1,4 @@
-import { LockEvent, TryAcquireResponse } from "../locks/lock-pool";
-import { ReleaseFunction } from "./release-function";
+import { TryAcquireResponse, LockEvent } from "../types/lock-options.type";
 
 export interface Lock {
   acquire(): Promise<ReleaseFunction>;
@@ -8,4 +7,17 @@ export interface Lock {
   releaseAcquiredLocks(): void;
   on(event: LockEvent, listener: (...args: any[]) => void): Lock;
   off(event: LockEvent, listener: (...args: any[]) => void): Lock;
+}
+
+export interface WaitingLock {
+  resolve(result: ReleaseFunction): void;
+  reject(reason?: any): void;
+}
+
+export interface AcquiredLock {
+  releaseTimeoutId?: NodeJS.Timeout;
+}
+
+export interface ReleaseFunction {
+  (): void;
 }
