--- src/interfaces/release-function.ts
@@ -1,3 +1,0 @@
-export interface ReleaseFunction {
-  (): void;
-}
