+++ src/interfaces/task.executor.ts
@@ -1,6 +1,4 @@
-import { TaskEvent, TryRunResponse } from "../task-executors/promise-pool";
-import { TaskOptions } from "../types/promise-options.type";
-import { TaskEntry } from "./task-entry";
+import { DiscardReason, ReleaseBeforeFinishReason, TaskEvent, TaskOptions, TryRunResponse } from "../types/promise-options.type";
 
 export interface TaskExecutor<T> {
   run<T>(task: (arg?: any) => Promise<T>, arg?: any, options?: TaskOptions): Promise<PromiseSettledResult<T>>;
@@ -15,4 +13,30 @@ export interface TaskExecutor<T> {
   waitingTasks(): number;
   runningTasks(): number;
   expiredTasks(): number;
+}
+
+export interface TaskEntry {
+  arg?: any | undefined;
+  options?: TaskOptions | undefined;
+  releaseReason?: ReleaseBeforeFinishReason;
+  discardReason?: DiscardReason;
+}
+
+export interface WaitingTask extends TaskEntry {
+  resolve(result: AcquireResponse): void;
+  reject(reason?: any): void;
+  waitingTimeoutId?: NodeJS.Timeout;
+}
+
+export interface RunningTask extends TaskEntry {
+  releaseTimeoutId?: NodeJS.Timeout;
+}
+
+export interface TaskExecutorReleaseFunction {
+  (reason?: ReleaseBeforeFinishReason): void;
+}
+
+export interface AcquireResponse {
+  release: TaskExecutorReleaseFunction;
+  taskEntry: TaskEntry;
 }
