+++ src/locks/lock-mutex.ts
@@ -1,8 +1,5 @@
-import { QueueType } from "../types/queue.type";
-import { TimeoutHandler } from "../types/timeout-handler.type";
-import { LockPool, PoolLockOptions } from "./lock-pool";
-
-export type LockOptions = { queueType?: QueueType; releaseTimeout?: number; releaseTimeoutHandler?: TimeoutHandler };
+import { LockPool } from "./lock-pool";
+import { LockOptions, PoolLockOptions } from "../types/lock-options.type";
 
 export class LockMutex extends LockPool {
   constructor(options?: LockOptions) {
