--- src/types/base-options.type.ts
@@ -1,1 +1,0 @@
-export class BaseOptions{}
