+++ src/types/promise-multi-step-options.type.ts
@@ -1,0 +1,4 @@
+import { Lock } from "../interfaces/lock";
+
+export type PromiseMultiStepOptions = { stepConcurrentLimits: number[] };
+export type MultiStepTask<T> = (...stepLocks: Lock[]) => Promise<T>;
