--- src/types/queue.type.ts
@@ -1,1 +1,0 @@
-export type QueueType = "FIFO" | "LIFO";
