--- src/types/task-timeout-handler.type.ts
@@ -1,3 +1,0 @@
-import { TaskEntry } from "../interfaces/task-entry";
-
-export type TaskTimeoutHandler = (taskEntry: TaskEntry) => void;
