--- src/types/timeout-handler.type.ts
@@ -1,1 +1,0 @@
-export type TimeoutHandler = () => void;
